[tornado] rename settings in constants
